refactor/comment: 엔티티 생성자 private 설정, Service 세분화
