+++ src/main/java/min/taskflow/auth/service/command/ExternalCommandAuthService.java
@@ -1,4 +1,4 @@
-package min.taskflow.auth.service.commandService;
+package min.taskflow.auth.service.command;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
