+++ src/main/java/min/taskflow/auth/service/query/ExternalQueryAuthService.java
@@ -1,0 +1,4 @@
+package min.taskflow.auth.service.query;
+
+public class ExternalQueryAuthService {
+}
