+++ src/main/java/min/taskflow/auth/service/query/InternalQueryAuthService.java
@@ -1,0 +1,4 @@
+package min.taskflow.auth.service.query;
+
+public class InternalQueryAuthService {
+}
