--- src/main/java/min/taskflow/auth/service/queryService/QueryExternalAuthService.java
@@ -1,4 +1,0 @@
-package min.taskflow.auth.service.queryService;
-
-public class QueryExternalAuthService {
-}
