+++ src/main/java/min/taskflow/comment/service/command/InternalCommandCommentService.java
@@ -1,0 +1,4 @@
+package min.taskflow.comment.service.command;
+
+public class InternalCommandCommentService {
+}
