+++ src/main/java/min/taskflow/comment/service/query/InternalQueryCommentService.java
@@ -1,0 +1,4 @@
+package min.taskflow.comment.service.query;
+
+public class InternalQueryCommentService {
+}
