+++ src/test/java/min/taskflow/auth/RefreshTokenTest.java
@@ -3,7 +3,7 @@ package min.taskflow.auth;
 import min.taskflow.auth.exception.AuthErrorCode;
 import min.taskflow.auth.exception.AuthException;
 import min.taskflow.auth.jwt.JwtUtil;
-import min.taskflow.auth.service.commandService.ExternalCommandAuthService;
+import min.taskflow.auth.service.command.ExternalCommandAuthService;
 import min.taskflow.user.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -40,7 +40,7 @@ class RefreshTokenTest {
                 .isInstanceOf(AuthException.class)
                 .extracting("errorCode")
                 .isEqualTo(AuthErrorCode.INVALID_REFRESH_TOKEN);
-        
+
         verify(userRepository, never()).findById(any());
         verify(jwtUtil, never()).createAccessToken(any(), any());
     }
